Fixed issues and added documentation
