+++ .vscode/launch.json
@@ -5,6 +5,12 @@
     "version": "0.2.0",
     "configurations": [
         {
+            "type": "node",
+            "request": "launch",
+            "name": "Launch Program",
+            "program": "${file}"
+        },
+        {
             "type": "extendscript-debug",
             "request": "launch",
             "name": "Run current file",
