+++ index.html
@@ -30,9 +30,10 @@
 
 			// For functions which require interaction at the JavaScript level, we provide these JQuery-based
 			// handlers, instead of directly invoking ExtendScript .This givs the JavaScript layer a chance
-			// to pass data into the ExtendScript layer, and process the results.
-
-
+            // to pass data into the ExtendScript layer, and process the results.
+            
+            // Start the Express Host
+            run();
 		});
 	</script>
 </head>
