--- localServer.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-	<meta charset="utf-8">
-	<script>
-		/* This script uses cep_node to start the Node.js server located at '/server/main.js' */
-		alert(__dirname + 'main.js');
-	    var localServer = cep_node.require(__dirname + 'main.js')();
-	</script>
-	<title>Import Example App</title>
-</head>
-<body>
-</body>
-</html>
