+++ package.json
@@ -1,8 +1,8 @@
 {
-  "name": "com.dnv.server",
+  "name": "com.dnv.harnesspanel",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "main.js",
   "dependencies": {
     "body-parser": "^1.18.2",
     "express": "^4.16.3",
